requirements.txt
selenium>=4.0
beautifulsoup4>=4.13
requests>=2.32
python-dotenv>=1.0
supabase>=2.16
